slight change

+++ Own/Bsp/DWT/SuperDWT.hpp
@@ -19,7 +19,7 @@ public:
     static void delay_us(uint32_t u_delay_time) {
         uint32_t start     = DWT->CYCCNT;
         uint32_t delay_cnt = u_delay_time * (SystemCoreClock / 1000000);
-        while (static_cast<uint32_t>(DWT->CYCCNT - start) < delay_cnt);
+        while (DWT->CYCCNT - start < delay_cnt);
     };
 
 private:
